Update GitHub workflow for release creation & Version bump to V0.3.5
The updates in the GitHub actions workflow are related to the release creation process. The key change is the setting of the 'prerelease' flag from 'true' to 'false'. This indicates that the releases will no longer be considered as prereleases in the repository, but will be treated as full-fledged product releases. This adjustment aligns with our product development stage, as it is now mature and stable enough for deployment in production environments.