Add 'diff-output.txt' to .gitignore
With this modification, we have decided to include 'diff-output.txt' in the .gitignore file. We no longer want to track this file in our version control as it mainly includes temporary output used during the debugging and development process. Leaving it out will help keep our repository clean and focused on source code files.