Update release note generation in CreateRelease workflow
Changed the method of generating release notes in the GitHub Actions workflow. Previously, the notes were directly assigned from the `git log` command to an environment variable. This has been altered to write the release notes into a temporary file first, then read and assign them to the environment variable from this file. This approach provides a better tracing of the release notes creation process and easier debugging in case of any issues.