Refactor GitHub release process in workflow script
This update refactors the way we generate release notes and create GitHub releases within our CI/CD workflow. Instead of generating release notes and storing them in a temporary text file, we are now immediately assigning the notes to an environment variable for subsequent use with the GitHub actions library. This simplifies our approach and removes the need for creating, reading, and disposing of the temporary file.

Additionally, we've shifted from using the generic 'github-script' action to the more dedicated 'create-release' and 'upload-release-asset' actions for creating the GitHub release and uploading the bundled file respectively. This modification makes the workflow more readable, maintainable, and less prone to potential issues arising from using 'github-script'. Furthermore, it enhances the relevancy and descriptiveness of the GitHub actions used.