Update GitHub Actions versions used in workflow
In this commit, we've upgraded the versions of our actions to ensure we're up to date with the latest features and bug fixes. Specifically, we updated the 'checkout' and 'setup-node' GitHub Actions used in our CreateRelease workflow from v2 to v3. As part of these updates, we maintain the current configuration values (checkout action fetch-depth set to 0, and Node.js version defined as '20.8.0'). These changes aim to keep our CI/CD process as effective and efficient as possible.