Update Node.js setup action version in GitHub workflow
We updated the version of the GitHub action used for setting up Node.js in our release creation workflow. This is in line with keeping our infrastructure up-to-date and leveraging the latest features and bug fixes provided by the newer version. It's important to note that this change could potentially affect the building and deployment processes, although it is not expected to cause any disruptive effects.