Update Typescript ESLint plugin and parser, esbuild versions
To keep up with bug-fixes and new feature updates, it was necessary to upgrade the versions of some dependencies. The "@typescript-eslint/eslint-plugin" and "@typescript-eslint/parser" versions were updated to 7.10.0. This allows us to take advantage of the latest ESLint rules for Typescript. The "esbuild" version was updated to 0.21.3. The increased version number for "esbuild" enhances our JavaScript build speed and efficiency. No breaking changes are expected as a result of these updates.