Add Jest for testing and update project metadata
This commit introduces Jest as the testing framework for the project. Configuration for Jest has been added, and the 'jest' and '@types/jest' packages have been added to the devDependencies. Several new scripts for running tests and modifying test behavior have also been added to 'package.json'.

Moreover, this update changes the name of the project in 'package.json' to 'obsidian-prj-plugin'.

These adjustments are an essential step in developing a robust project with reliable and maintainable code, by enabling more efficient testing and accurate project identification.