Refactor search functionality in rendering classes
This commit brings a change in how search functionality is implemented in 'Document', 'Note' and 'Project' rendering classes. We moved from using static search to the instantiation of the Search class and the use of instance methods. This change provides more modular and maintainable structure. Simultaneously, it accounts for refining the data conversion to string for search logic application, thus providing further robustness. Please note that it involves changes in the search property in BlockRenderSettings, now pointing to a Search instance instead of a SearchTermsArray, hence, follow-up adaptions might be necessary.