Refactoring/search (#40)
* Add Jest for testing and update project metadata

This commit introduces Jest as the testing framework for the project. Configuration for Jest has been added, and the 'jest' and '@types/jest' packages have been added to the devDependencies. Several new scripts for running tests and modifying test behavior have also been added to 'package.json'.

Moreover, this update changes the name of the project in 'package.json' to 'obsidian-prj-plugin'.

These adjustments are an essential step in developing a robust project with reliable and maintainable code, by enabling more efficient testing and accurate project identification.

* Add testing phase into GitHub release workflow

In an effort to ensure code quality and functionality for each release, a new step to run tests has been incorporated into the existing release GitHub action workflow. The "Run Tests" step uses the 'npm run test:verbose' command for an expressive test execution. This ensures that any issues are recognized and resolved prior to the build process, enhancing stability for the final release.

* Add performance and unit tests for SearchParser and SearchQuery

This commit introduces comprehensive unit tests for 'SearchParser' and 'SearchQuery' to ensure correct functionality for various search queries. It evaluates single term parsing, multi-term parsing, parsing with different operators, negated terms, quoted terms, and even complex queries. Error handling is also tested, for cases like unmatched quotes and consecutive operators.

* Introduce search functionality in application

The commit adds advanced search capability enabling users to enter a search query with terms, operators (AND, OR, NOT), and quotes. The search logic parses the query to form individual text matches and operates according to the specified criteria, returning results that adhere to the defined parameters.

Introduced classes include:
- `SearchParser` for breaking down user's search string into components.
- `SearchQuery`, `SearchElement`, `SearchTerm` and `SearchOperator` to model the search and its elements.
- `Search` to apply the logic of the search.

This development is a response to the need for granular search function in the application, allowing users to define precise or complex queries for more targeted results. Please verify proper error handling and correct parsing of user inputs.

* Subject: Deprecate old Search methods

Two methods in the `Search` class - `parseSearchText` and `applySearchLogic` - have been marked deprecated. These methods are set to be removed in the future and plans are to replace them with more efficient code through the new `Search` class to improve the search functionality.

* Refactor tag extraction from active file

The logic for extracting tags from the active file has been shifted from the BaseModalForm to a more appropriate Tags class. This allows for better organization of code, and a more logical delegation of responsibilities, adhering to the Single Responsibility Principle. No significant changes in functionality are expected from this update.

* Refactor search functionality in rendering classes

This commit brings a change in how search functionality is implemented in 'Document', 'Note' and 'Project' rendering classes. We moved from using static search to the instantiation of the Search class and the use of instance methods. This change provides more modular and maintainable structure. Simultaneously, it accounts for refining the data conversion to string for search logic application, thus providing further robustness. Please note that it involves changes in the search property in BlockRenderSettings, now pointing to a Search instance instead of a SearchTermsArray, hence, follow-up adaptions might be necessary.

* Version bump to V0.4.0