Refactor search and filter logic into base class
Moved the search and filter methods from the derived classes into the 'TableBlockRenderComponent' base class to eliminate code duplication. This change involved shifting the 'onSearch' and 'onFilter' methods, along with supporting logic, up from the 'DocumentBlockRenderComponent', 'NoteBlockRenderComponent', and 'ProjectBlockRenderComponent' subclasses. It has also led to the necessity of changing 'getHideState' to an abstract method in the base class. These modifications make the codebase easier to manage and understand, improving maintainability. Furthermore, they ensure a unified flow of search and filter operations across the different classes at a higher level.