Remove deprecated Search class and related types
The Search class, including its methods 'parseSearchText' and 'applySearchLogic', have been deleted, as they are no longer in use. They were marked as deprecated in favor of a new Search class. This commit also removes the types 'SearchTerm' and 'SearchTermsArray' which were specific to the deprecated class. The removal of these unused entities simplifies the codebase and reduces maintenance overhead.