Add `first` and `last` methods to `Tags` class
Introduced two new methods, `first` and `last`, to the `Tags` class. These methods return the first and last tag in the array, respectively. If the tags array is empty, they return `undefined`. This enhancement improves the accessibility of tag retrieval in the system. Corresponding unit tests were also added to ensure both methods are functioning as expected.